Ensure we have at least 2 Conveyor::Job subclasses
The Crystal type checker crashes if there aren't at least 2 Job
subclasses. I think this is because it's specializing on the single job
class available and trying to downcast. I want to say it's a known issue
in Crystal because this isn't the first time I've come across it, but I
can't say for sure.

We expose one of the job classes because it turns out I actually do run
a lot of empty jobs when testing edge cases and things, so it'd be nice
to just provide one since it's already there.